refactor(typescript): move typescript code to src/ folder and generate output in lib/ folder
